backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-dotenv==1.0.0
requests==2.31.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0